system/transcriber/requirements.txt
fastapi
uvicorn
transformers
soundfile
librosa
python-multipart
boto3
python-dotenv
kafka-python
pyannote.audio
ffmpeg-python
pydub
psycopg2